Added type module into index, also data js, util.js files and export/import between them
